% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caMST.R
\name{caMSTStartup}
\alias{caMSTStartup}
\title{Create Package Startup Message}
\usage{
caMSTStartup()
}
\description{
Makes package startup message.
}
\details{
Idea taken from https://github.com/ntguardian/MCHT/blob/master/R/StartupMessage.R
}
\examples{
caMST:::caMSTStartup()
}
