% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caMST.R
\name{.onAttach}
\alias{.onAttach}
\title{Package Attach Hook Function}
\usage{
.onAttach(lib, pkg)
}
\arguments{
\item{lib}{a character string giving the library directory where the package
defining the namespace was found}

\item{pkg}{a character string giving the name of the package}
}
\description{
Hook triggered when package attached.
}
\details{
Idea taken from https://github.com/ntguardian/MCHT/blob/master/R/StartupMessage.R
}
\examples{
caMST:::.onAttach(.libPaths()[1], "caMST")
}
