% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_module_items}
\alias{example_module_items}
\title{Example "item-to-module" map matrix, showcasing how the items and modules are related.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 42 rows and 7 columns.
}
\usage{
example_module_items
}
\description{
A matrix with items on the rows and modules on the columns, where 0 indicates
the item and module are unrelated and 1 indicates that the item is a part of
that module. Used in combination with a transition matrix to describe a
multistage adaptive test.
}
\keyword{datasets}
