% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_responses}
\alias{example_responses}
\title{Responses to all of the example items by the five individuals represented in
the "example_thetas" data.}
\format{An object of class \code{data.frame} with 5 rows and 598 columns.}
\usage{
example_responses
}
\description{
A data frame with individuals on the rows and items on the columns. The
values of the data frame are the response patterns of the individuals to all
of the items in the example item files.
}
\keyword{datasets}
