% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_thetas}
\alias{example_thetas}
\title{Theta values used in the examples.}
\format{An object of class \code{numeric} of length 5.}
\usage{
example_thetas
}
\description{
A numeric vector with five simulated theta values.
}
\keyword{datasets}
