% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_transition_matrix}
\alias{example_transition_matrix}
\title{Example transition matrix showing how individuals traverse the multistage test.}
\format{An object of class \code{matrix} with 7 rows and 7 columns.}
\usage{
example_transition_matrix
}
\description{
A matrix with modules on the rows and columns. A 0 indicates that an individual
 cannot move from the row module to the column module, while a 1 indicates that
 an individual who has completed the row module can potentially transition
 into the column module.
}
\keyword{datasets}
