% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mst_only_matrix}
\alias{mst_only_matrix}
\title{Example "item-to-module" map matrix for the "multistage_test" example.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 42 rows and 7 columns.
}
\usage{
mst_only_matrix
}
\description{
A matrix with items on the rows and modules on the columns, where 0 indicates
the item and module are unrelated and 1 indicates that the item is a part of
that module. Used in combination with a transition matrix to describe a
multistage adaptive test.
}
\keyword{datasets}
