% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caMST.R
\docType{package}
\name{caMST}
\alias{caMST}
\alias{caMST-package}
\title{\code{caMST} package}
\description{
Computer Adaptive Mutistage Test Analysis
}
\details{
See the README on \href{https://github.com/AnthonyRaborn/caMST#readme}{GitHub} for more information.
}
