% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat_items}
\alias{cat_items}
\title{Example items for the CAT stage of an example adaptive test.}
\format{
A data frame with 564 rows (items) and 6 columns (item parameters):
\describe{
\item{a}{item discrimination}
\item{b}{item difficulty}
\item{c}{item guessing}
\item{u}{item carelessness}
\item{content_ID}{what content area the item comes from}
\item{stage}{which stage the item belongs to}
}
}
\usage{
cat_items
}
\description{
A data frame formatted in the style required by \pkg{mstR} for item data.
Represents items to be used in an item-level adaptation portion of a
computerized adaptive test.
}
\keyword{datasets}
