% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computerized_adaptive_test_function.R
\name{computerized_adaptive_test}
\alias{computerized_adaptive_test}
\title{Computerized Adaptive Test}
\usage{
computerized_adaptive_test(cat_item_bank, response_matrix, initial_theta = 0,
  model = NULL, randomesque = 1, maxItems = 50, method = "BM", ...)
}
\arguments{
\item{cat_item_bank}{A data frame with the items on the rows and their item parameters on the columns. These should be in the \pkg{catR} package format for item banks.}

\item{response_matrix}{A matrix of the person responses, with individuals as rows and items as columns.}

\item{initial_theta}{The initial theta estimate for all individuals. Default is 0.}

\item{model}{Either NULL (default) for dichotomous models or a character value indicating the polytomous model used. See the\pkg{mstR} package for more details.}

\item{randomesque}{An integer value that indicates the number of items from which the selection rule should choose from randomly for administration. See the help documentation for \code{catR::nextItem} for more details.}

\item{maxItems}{An integer value indicating the maximum number of items to administer, regardless of other stopping rules.}

\item{method}{A character value indicating method for the provisional theta estimate. Defaults to "BM" (Bayes Modal). See the \pkg{catR} package for more details.}

\item{...}{Further arguments to be passed to \code{catR::nextItem}. Currently unimplemented.}
}
\value{
List of results.
}
\description{
Computerized Adaptive Test
}
\examples{
#fill in later
}
