% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class_generics.R
\docType{class}
\name{MAT-class}
\alias{MAT-class}
\title{An S4 method for mixed adaptive tests.}
\value{
An S4 object of class `MAT`.
}
\description{
An S4 method for mixed adaptive tests.
}
\section{Slots}{

\describe{
\item{\code{function.call}}{The original function call.}

\item{\code{final.theta.estimate.catR}}{Numeric vector of theta estimates calculated by the provided `method`.}

\item{\code{eap.theta}}{Numeric vector of theta estimates calculated by `catR::eapEst`.}

\item{\code{final.theta.Baker}}{Numeric vector of theta estimates calculated by the internal `iterative.theta.estimate` function.}

\item{\code{final.theta.SEM}}{Numeric vector of SEM estimates calculated by the internal `iterative.theta.estimate` function.}

\item{\code{final.items.seen}}{Character matrix of the final items seen by each individual.}

\item{\code{modules.seen}}{Numeric matrix of the modules seen by each individual.}

\item{\code{final.responses}}{Numeric matrix of the response patterns observed.}

\item{\code{transition.matrix}}{Numeric matrix; the transition matrix entered into the function.}

\item{\code{n.stages}}{Numeric; the number of stages specified.}

\item{\code{runtime}}{A `difftime` object of the total run time of the function.}
}}

