% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_adaptive_test_function.R
\name{mixed_adaptive_test}
\alias{mixed_adaptive_test}
\title{Mixed Computerized Adaptive Multistage Test}
\usage{
mixed_adaptive_test(response_matrix, cat_item_bank, initial_theta = 0,
  method = "BM", item_method = "MFI", cat_length, nAvailable_cat = NULL,
  cbControl = NULL, cbGroup = NULL, randomesque = 1, mst_item_bank,
  modules)
}
\arguments{
\item{response_matrix}{A matrix of the person responses, with individuals as rows and items as columns.}

\item{cat_item_bank}{A data frame with the first stage items on the rows and their item parameters on the columns. These should be in the \pkg{mstR} package format for item banks.}

\item{initial_theta}{The initial theta estimate for all individuals.}

\item{method}{A character value indicating method for the provisional theta estimate. Defaults to "BM" (Bayes Modal). See the \pkg{catR} and \pkg{mstR} packages for more details.}

\item{item_method}{A character value indicating the method for the item-level selection in the first stage. Defaults to "MFI" (Maximum Fisher Information). See the \pkg{catR} and \pkg{mstR} packages for more details.}

\item{cat_length}{A numeric value indicating the number of items in the first stage.}

\item{nAvailable_cat}{Defaults to `NULL`. See the \pkg{catR} package for more information on how to use this option.}

\item{cbControl}{A list of the appropriate format used to control for content balancing in the first stage. See the Details in the nextItem function in \pkg{catR}.}

\item{cbGroup}{A factor vector of the appropriate format used to control for content balancing in the first stage. See the Details in the nextItem function in \pkg{catR}.}

\item{randomesque}{An integer indicating the number of items from which to select the next item to administer in the first stage. Default value is 1.}

\item{mst_item_bank}{A data frame with the second stage and beyond items on the rows and their item parameters on the columns. These should be in the \pkg{mstR} package format for item banks.}

\item{modules}{A matrix describing the relationship between the items and the modules they belong to. See \strong{Details}.}
}
\description{
Mixed Computerized Adaptive Multistage Test
}
\details{

}
\examples{

}
