% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix_plot.R
\name{transition_matrix_plot}
\alias{transition_matrix_plot}
\title{Transition Matrix Plot}
\usage{
transition_matrix_plot(transition_matrix = NULL, n_each_stage = NULL)
}
\arguments{
\item{transition_matrix}{A matrix describing how individuals can transition from one stage to the next.}

\item{n_each_stage}{A numeric vector indicating how many modules exist at each stage.}
}
\value{
A plot using the current graphic device.
}
\description{
Given a transition matrix and the number of modules at each stage, produces a plot
that demonstrates the potential paths through a (mixed) multistage test.
}
\examples{
# Create a plot for a multistage test with a 1-3-3 design
data('example_transition_matrix')
transition_matrix_plot(example_transition_matrix, n_each_stage = c(1,3,3))

\donttest{
# Save the plot as a png file.
png("Example 1-3-3 Transition Matrix Plot.png")
transition_matrix_plot(example_transition_matrix, n_each_stage = c(1,3,3))
title("Transition Matrix for a 1-3-3 Design MST")
dev.off()
}

}
